+++ lib/presentation/pages/splash/splash_page.dart
@@ -1,0 +1,24 @@
+import 'package:flutter/material.dart';
+
+class SplashPage extends StatelessWidget{
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      body: Center(
+      child: Column(
+        mainAxisAlignment: MainAxisAlignment.center,
+        children: [
+        Text('SCDream폰트 확인용입니다. SCDream폰트 확인용입니다.', style: TextStyle(fontWeight: FontWeight.w100),),
+        Text('SCDream폰트 확인용입니다. SCDream폰트 확인용입니다.', style: TextStyle(fontWeight: FontWeight.w200),),
+        Text('SCDream폰트 확인용입니다. SCDream폰트 확인용입니다.', style: TextStyle(fontWeight: FontWeight.w300),),
+        Text('SCDream폰트 확인용입니다. SCDream폰트 확인용입니다.', style: TextStyle(fontWeight: FontWeight.w400),),
+        Text('SCDream폰트 확인용입니다. SCDream폰트 확인용입니다.', style: TextStyle(fontWeight: FontWeight.w500),),
+        Text('SCDream폰트 확인용입니다. SCDream폰트 확인용입니다.', style: TextStyle(fontWeight: FontWeight.w600),),
+        Text('SCDream폰트 확인용입니다. SCDream폰트 확인용입니다.', style: TextStyle(fontWeight: FontWeight.w700),),
+        Text('SCDream폰트 확인용입니다. SCDream폰트 확인용입니다.', style: TextStyle(fontWeight: FontWeight.w800),),
+        Text('SCDream폰트 확인용입니다. SCDream폰트 확인용입니다.', style: TextStyle(fontWeight: FontWeight.w900),),
+      ],),
+    ),
+    );
+  }
+}
